file_converter/requirements.txt
# file_converter/requirements.txt
# 轻量级文件转换依赖

# Web 框架
fastapi==0.104.1
uvicorn[standard]==0.24.0
python-multipart==0.0.6

# 文件转换库
pdfplumber==0.10.3          # PDF 文本/表格提取
python-pptx==0.6.23         # PPTX 解析
python-docx==1.1.0          # DOCX 解析
openpyxl==3.1.2             # Excel 解析

# OCR 依赖
pytesseract==0.3.10         # Tesseract Python 封装
pdf2image==1.16.3           # PDF 转图片 (用于扫描件 OCR)
Pillow>=10.0.0              # 图像处理
requests>=2.31.0            # 语言包下载

# 测试依赖（可选，仅开发环境）
# pytest==7.4.3
# httpx==0.25.2
